+++ pydl/pydlutils/sdss/sdss_sweep_circle.py
@@ -53,27 +53,27 @@ def sdss_sweep_circle(ra,dec,radius,stype='star',allobj=False):
     #
     ira = np.array([ra])
     idec = np.array([dec])
-    m1,m2,d12 = spherematch(ira,idec,index['RA'],index['DEC'],radius+0.36,maxmatch=0)
-    if len(m2) == 0:
+    m1,m2,d12 = spherematch(index['RA'],index['DEC'],ira,idec,radius+0.36,maxmatch=0)
+    if len(m1) == 0:
         return None
     if not allobj:
-        w = index['NPRIMARY'][m2] > 0
+        w = index['NPRIMARY'][m1] > 0
         if w.any():
-            m2 = m2[w]
+            m1 = m1[w]
         else:
             return None
     #
     # Maximum number of objects
     #
     if allobj:
-        n = index['IEND'][m2] - index['ISTART'][m2] + 1
+        n = index['IEND'][m1] - index['ISTART'][m1] + 1
         ntot = (np.where(n>0,n,np.zeros(n.shape,dtype=n.dtype))).sum()
     else:
-        ntot = index['NPRIMARY'][m2].sum()
+        ntot = index['NPRIMARY'][m1].sum()
     #
     # Find unique run + camcol
     #
-    rc = index['RUN'][m2]*6 + index['CAMCOL'][m2] - 1
+    rc = index['RUN'][m1]*6 + index['CAMCOL'][m1] - 1
     isort = rc.argsort()
     iuniq = uniq(rc[isort])
     istart = 0
@@ -85,10 +85,10 @@ def sdss_sweep_circle(ra,dec,radius,stype='star',allobj=False):
         #
         # Determine which file and range of rows
         #
-        run = index['RUN'][m2[icurr[0]]]
-        camcol = index['CAMCOL'][m2[icurr[0]]]
-        rerun = index['RERUN'][m2[icurr[0]]]
-        fields = index[m2[icurr]]
+        run = index['RUN'][m1[icurr[0]]]
+        camcol = index['CAMCOL'][m1[icurr[0]]]
+        rerun = index['RERUN'][m1[icurr[0]]]
+        fields = index[m1[icurr]]
         ist = fields['ISTART'].min()
         ind = fields['IEND'].max()
         if ind >= ist:
@@ -103,9 +103,9 @@ def sdss_sweep_circle(ra,dec,radius,stype='star',allobj=False):
                 #
                 # Keep only objects within the desired radius
                 #
-                tm1,tm2,d12 = spherematch(ira,idec,tmp_objs['RA'],tmp_objs['DEC'],radius,maxmatch=0)
-                if len(tm2) > 0:
-                    tmp_objs=tmp_objs[tm2]
+                tm1,tm2,d12 = spherematch(tmp_objs['RA'],tmp_objs['DEC'],ira,idec,radius,maxmatch=0)
+                if len(tm1) > 0:
+                    tmp_objs=tmp_objs[tm1]
                     #
                     # Keep only SURVEY_PRIMARY objects by default
                     #